Adding file specific embargo check
Adding a check for file specific embargo.  It opens the header of each file before download and checks whether type is 'text/html'.  If it is html, it looks for the phrase "<title>University Digital Conservancy :: Login" (which indicates that it has redirected to a UDC log-in page) and does not try to download the file.  If it is not html or does not include that phrase, it proceeds with download.  This commit comments out (but does not remove) the previous attempt to identify embargos using accessStatus.